refactor: abstract MonacoEditor to allow for shared props
