+++ webui/react/src/components/MonacoEditor.module.scss
@@ -1,0 +1,3 @@
+.base {
+  height: 100%;
+}
