+++ webui/react/src/pages/ExperimentDetails/ExperimentConfiguration.module.scss
@@ -1,3 +1,3 @@
-.base {
-  height: 100%;
+.loading {
+  padding: var(--theme-sizes-layout-mega);
 }
