+++ webui/react/src/pages/ExperimentDetails/ExperimentConfiguration.tsx
@@ -1,9 +1,8 @@
 import yaml from 'js-yaml';
-import React, { useEffect, useRef } from 'react';
-import MonacoEditor from 'react-monaco-editor';
+import React from 'react';
 
 import Section from 'components/Section';
-import useResize from 'hooks/useResize';
+import Spinner from 'components/Spinner';
 import { ExperimentBase } from 'types';
 
 import css from './ExperimentConfiguration.module.scss';
@@ -12,35 +11,23 @@ interface Props {
   experiment: ExperimentBase;
 }
 
-const PADDING = 8;
+const MonacoEditor = React.lazy(() => import('components/MonacoEditor'));
 
 const ExperimentConfiguration: React.FC<Props> = ({ experiment }: Props) => {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const editorRef = useRef<MonacoEditor>(null);
-  const resize = useResize(containerRef);
-
-  useEffect(() => {
-    editorRef.current?.editor?.layout();
-  }, [ resize ]);
-
   return (
     <Section bodyNoPadding bodyScroll maxHeight>
-      <div className={css.base} ref={containerRef}>
+      <React.Suspense fallback={(
+        <div className={css.loading}><Spinner tip="Loading text editor..." /></div>
+      )}>
         <MonacoEditor
           height="100%"
-          language="yaml"
           options={{
-            minimap: { enabled: false },
             occurrencesHighlight: false,
-            padding: { bottom: PADDING, top: PADDING },
             readOnly: true,
-            scrollBeyondLastLine: false,
-            selectOnLineNumbers: true,
           }}
-          ref={editorRef}
-          theme="vs-light"
-          value={yaml.dump(experiment.configRaw)} />
-      </div>
+          value={yaml.dump(experiment.configRaw)}
+        />
+      </React.Suspense>
     </Section>
   );
 };
